requirements.txt
discord.py==1.5.1
google-api-python-client==1.12.8